Light refactor
- use setup() throughout in tests
- Consistently use core over unknown
