+++ internal/core_test.go
